refactor: move init.py utils to utils/*
