+++ frappe/utils/__init__.py
@@ -1143,3 +1143,11 @@ class CallbackManager:
 
 	def reset(self):
 		self._functions.clear()
+
+
+def safe_eval(code, eval_globals=None, eval_locals=None):
+	"""A safer `eval`"""
+
+	from frappe.utils.safe_exec import safe_eval
+
+	return safe_eval(code, eval_globals, eval_locals)
